requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
pymysql
python-dotenv
pydantic-settings
httpx